Add Validation for user input
