+++ src/components/sidebar/sidebar.js
@@ -1,91 +1,7 @@
 // import projectsRender from '../projects/projects';
 import './sidebar.scss';
 import renderProjects from '../projects/projects';
-import Project from '../../utils/project';
-
-function renderForm(projects) {
-  const formContainer = document.createElement('div');
-  const form = document.createElement('form');
-  const header = document.createElement('h2');
-  const titleContainer = document.createElement('div');
-  const titleLabel = document.createElement('label');
-  const titleInput = document.createElement('input');
-  const descriptionContainer = document.createElement('div');
-  const descriptionLabel = document.createElement('label');
-  const descriptionInput = document.createElement('textarea');
-  const submitBtn = document.createElement('button');
-  const closeBtn = document.createElement('span');
-
-  formContainer.id = 'project-form';
-  titleInput.id = 'title';
-  descriptionInput.id = 'description';
-  submitBtn.id = 'project-submit';
-
-  formContainer.classList.add('form-container');
-  form.classList.add('position-relative');
-  titleContainer.className = 'form-group';
-  titleInput.className = 'form-control';
-  descriptionContainer.className = 'form-group';
-  descriptionInput.className = 'form-control';
-  submitBtn.classList.add('btn', 'btn-primary');
-  closeBtn.className = 'close-btn';
-
-  titleLabel.setAttribute('for', 'title');
-  titleInput.setAttribute('type', 'text');
-  titleInput.setAttribute('name', 'title');
-  descriptionLabel.setAttribute('for', 'description');
-  descriptionInput.setAttribute('name', 'description');
-  descriptionInput.setAttribute('rows', '5');
-  submitBtn.setAttribute('type', 'submit');
-
-  titleLabel.textContent = 'Title';
-  descriptionLabel.textContent = 'Description';
-  submitBtn.textContent = 'Submit';
-  header.textContent = 'New Project';
-
-  formContainer.appendChild(form);
-  form.appendChild(closeBtn);
-  form.appendChild(header);
-  form.appendChild(titleContainer);
-  titleContainer.appendChild(titleLabel);
-  titleContainer.appendChild(titleInput);
-  form.appendChild(descriptionContainer);
-  descriptionContainer.appendChild(descriptionLabel);
-  descriptionContainer.appendChild(descriptionInput);
-  form.appendChild(submitBtn);
-  document.body.appendChild(formContainer);
-
-  function removeForm2(e) {
-    if (e.key === 'Escape') {
-      formContainer.remove();
-    }
-
-    document.removeEventListener('keydown', removeForm2);
-  }
-
-  function removeForm(e) {
-    formContainer.remove();
-    e.target.removeEventListener('click', removeForm);
-    document.removeEventListener('keydown', removeForm2);
-  }
-
-  closeBtn.addEventListener('click', removeForm);
-  document.addEventListener('keydown', removeForm2);
-  submitBtn.addEventListener('click', (e) => {
-    e.preventDefault();
-
-    const title = titleInput.value;
-    const description = descriptionInput.value;
-
-    const newProject = new Project(title, description);
-    projects.push(newProject);
-    renderProjects(projects);
-
-    removeForm(e);
-
-    renderProjects(projects);
-  });
-}
+import renderForm from '../projectForm/form';
 
 export default function render(projects) {
   const { body } = document;
