+++ src/components/sidebar/sidebar.scss
@@ -51,7 +51,7 @@
 }
 
 .form-container form {
-  width: 400px;
+  width: 300px;
   background-color: white;
   padding: 30px;
 }
